refactor: Mean Squared Errors after redefining Losses interfaces
